# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Medicines ─────────────────────────────────────────────────────────────
...
  /purchase-orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchase-orders]
      summary: List purchase/restock orders
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create a restock order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchase-orders]
      summary: Get purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchase-orders/{id}/receive:
    patch:
      operationId: receivePurchaseOrder
      tags: [purchase-orders]
      summary: Mark purchase order as received (updates stock)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Reports ───────────────────────────────────────────────────────────────
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales summary — orders by day/range
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Revenue aggregated by date range
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string

    # ── Medicine ──────────────────────────────────────────────────────────
...
    PurchaseOrderItem:
      type: object
      required: [id, purchaseOrderId, medicineId, quantity, unitPrice]
      properties:
        id:
          type: integer
        purchaseOrderId:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: string
    PurchaseOrderItemInput:
      type: object
      required: [medicineId, quantity, unitPrice]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: string
    PurchaseOrder:
      type: object
      required: [id, supplierId, status, total, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, received, cancelled]
        total:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItem"
        createdAt:
          type: string
          format: date-time
    PurchaseOrderInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
          minItems: 1

    # ── Reports ───────────────────────────────────────────────────────────
    SalesByDay:
      type: object
      required: [date, orders, revenue]
      properties:
        date:
          type: string
        orders:
          type: integer
        revenue:
          type: string
    SalesReport:
      type: object
      required: [totalOrders, totalRevenue, byDay]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: string
        byDay:
          type: array
          items:
            $ref: "#/components/schemas/SalesByDay"
...
    RevenueByRange:
      type: object
      required: [date, revenue]
      properties:
        date:
          type: string
        revenue:
          type: string
    RevenueReport:
      type: object
      required: [totalRevenue, byDate]
      properties:
        totalRevenue:
          type: string
        byDate:
          type: array
          items:
            $ref: "#/components/schemas/RevenueByRange"