# lib/api-zod/src/generated/types/supplierUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplierUpdate {
  /** @minLength 1 */
  name?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  address?: string;
}